# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: BookMind Foundation API Specification (BM-PRD-02)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /books:
    get:
      operationId: getBooks
      tags: [books]
      summary: List user books
      description: Returns list of books in the user library with reading progress
      responses:
        "200":
          description: List of books
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /books/{bookId}:
    get:
      operationId: getBookDetails
      tags: [books]
      summary: Get book details
      description: Returns book metadata, pages, progress, bookmarks, and notes
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Book details with pages and annotations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDetailResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBook
      tags: [books]
      summary: Delete book
      description: Soft-deletes or removes book and stored original file
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Book successfully deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /books/{bookId}/progress:
    put:
      operationId: updateReadingProgress
      tags: [books]
      summary: Update reading progress
      description: Persists current page, percentage, and completion status
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReadingProgressRequest"
      responses:
        "200":
          description: Reading progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadingProgress"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /books/{bookId}/bookmarks:
    post:
      operationId: createBookmark
      tags: [books]
      summary: Create bookmark
      description: Creates a bookmark on a specific page
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookmarkRequest"
      responses:
        "201":
          description: Bookmark created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /books/{bookId}/bookmarks/{bookmarkId}:
    delete:
      operationId: deleteBookmark
      tags: [books]
      summary: Delete bookmark
      description: Removes a bookmark by ID
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
        - name: bookmarkId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bookmark deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Bookmark not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /books/{bookId}/notes:
    post:
      operationId: createNote
      tags: [books]
      summary: Create note
      description: Creates a page note or quote highlight
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /books/{bookId}/notes/{noteId}:
    put:
      operationId: updateNote
      tags: [books]
      summary: Update note
      description: Edits note content or color tag
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
        - name: noteId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteRequest"
      responses:
        "200":
          description: Note updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Note not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteNote
      tags: [books]
      summary: Delete note
      description: Deletes note by ID
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
        - name: noteId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Note deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Note not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /books/import:
    post:
      operationId: importBookPdf
      tags: [books]
      summary: Import PDF book
      description: Uploads and ingests a PDF file into private user storage and creates a library book
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImportBookPdfRequest"
      responses:
        "201":
          description: Book successfully imported and ready for processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportBookResponse"
        "400":
          description: Invalid PDF format or validation failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Duplicate PDF already in library
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: PDF exceeds maximum size
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Corrupted or password protected PDF
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /books/{bookId}/processing:
    get:
      operationId: getBookProcessingStatus
      tags: [books]
      summary: Get book processing status
      description: Returns the processing job state and progress for a book
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current processing status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessingJobStatus"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Book not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          $ref: "#/components/schemas/ErrorDetail"
      required:
        - success
        - error
    BookSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        author:
          type: string
        totalPages:
          type: integer
        currentPage:
          type: integer
        progressPercent:
          type: integer
        coverUrl:
          type: string
        sourceType:
          type: string
        processingStatus:
          type: string
        lastReadAt:
          type: string
      required:
        - id
        - title
        - totalPages
        - currentPage
        - progressPercent
...
    UpdateReadingProgressRequest:
      type: object
      properties:
        currentPage:
          type: integer
          minimum: 1
        progressPercent:
          type: integer
          minimum: 0
          maximum: 100
        completed:
          type: boolean
      required:
        - currentPage
...
    CreateBookmarkRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          minimum: 1
        title:
          type: string
      required:
        - pageNumber
...
    BookDetailResponse:
      type: object
      properties:
        book:
          $ref: "#/components/schemas/BookSummary"
        pages:
          type: array
          items:
            $ref: "#/components/schemas/BookPage"
        progress:
          $ref: "#/components/schemas/ReadingProgress"
        bookmarks:
          type: array
          items:
            $ref: "#/components/schemas/Bookmark"
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
      required:
        - book
        - pages
        - bookmarks
        - notes
...